Parameter PMF reader
For generation interval, delays, and right-truncation.

It assumes that:
- We want a PMF
- The PMF is coming from a file with schema specified as in
  cdcent/cfa-parameter-estimates#9
- The parameter names and disease names are following a specified schema
- The file is in parquet format
- The PMFs are actually proper PMFs

It relaxes the assumptions that:
- The PMFs are coming from the same file
- The PMFs must be present (can skip delays and right-truncation but
  not GI)
- The files are in Azure or have a specific name/path

Unit tests cover successfully reading all the parameters individually
as well as in combination. They also check for failure in the expected
places for desired failure modes.

Switching over to this schema from manual CSVs now is a bit of a choice.
I took three cracks at this PR before landing on doing it this way. I
think it provides a couple of important benefits to make the switch now
and that comes through in designing the code here:
1. Our existing CSV-based approach produces 3 distinct files with
   close-ish schemas, but not an exact match. It requires distinct
   reader functions and substantially more code than reading from a
   file with a unified schema like I do here.
2. I think this approach helps avoid issues like this week's production
   mishap. It allows us to mix and match files as needed, and we can
   point to a single drop-in file for testing if desired. We don't need
   to fiddle with the production environment.
3. I think we're going to need to make a switch on the parameter
   approach sooner or later. I wanted to bite the bullet and get it
   over with all at once. It's convenient that it helps make the code
   simpler, but I think making changes swiftly rather than dragging
   things out provides its own benefit too.