Additional small CI tweaks
In #22 we fixed /document by expanding the number of scopes that could
call the command. Here we do the same thing for /style.

The other change is from the default github token to a token that I've
provided the repo. This change allows the pushed commits with the
auto-styled or -documented code to re-start CI.