Read from new parameters schema
Update reader to expect a reference date column. This change was a bit
finicky because we want to ignore the column when reading a GI or delay
and use the column when reading right-truncation.

I excuted this logic with an additional if block, but it tripped the
cyclocomp linter. I thought about it a bit, but still think this is the
cleanest approach (as opposed to breaking out the right-truncation
reader as its own special case or doing a post-read pre-check filter).
So I moved the checking into its own helper and left the dedicated
reader logic in the main function.

I also added some additional CLI info text on what the right-truncation
filter is looking for and what was selected. I think it's helpful, but
open to feedback that it's too much.

Closes #70