Add metadata output (#105)
* Add metadata output

This should help us save useful metadata into the metadata file. The current code saved nothing.

* Use an empty string if NULL

* add missing comma

* forgot to change these from list to class

* realized that production_date is the wrong spot

It was originally put there by copilot, and somehow I missed that it was there the whole time. This should have happened in the Config class PR. Not sure how it got through there.