Strip out commit message parsing
Because the parsing relies on checking out the last commit in the
PR. The assumptions and logic to handle this checkout break when
the branch isn't a PR, but instead the branch is main. We need to
be able to run this workflow on main to automate deployment to Batch.