Workflow to close PR pool on close (#125)
* Workflow to close PR pool on close

This workflow _should_ clean up the associated Batch pool when closing
the PR.

It doesn't clean up the associated tagged image in ACR (that's harder)
or jobs that were run on the pool. It would be nice to auto-delete any
linked jobs, but it seems tricker than I'd want? I'll plan to revisit
that when moving this into a dedicated action outside this repo.

Closes #62

* Bump NEWS

* Update .github/workflows/delete-pool-on-pr-close.yaml

* Document non-standard flag

* Drop workflow_dispatch option

* Fix indenting

* Switch over to pre-canned az cli actions