Hotfeature: rerun feature (#222)
* add output_container to Config object

* handy way to pick the right output container name

* Use the correct output container in pipeline

* specify output container in config generation

* forgot news

* remove output container as arg to function

* use rlang::is_empty instead of rlang::is_null

is empty checks for null, and for empty vectors

* forgot to document

* Use the newly created storage containers

We would like to use a new prod storage container so that all of the outputs have the same formatting.

Also update the `test-batch` command to use the new container for testing.

* get rid of unused function

* add commands for re-run

* swap from state_abb to state

* add news item

* add ability to specify report date when running

The default report date will be today. It is settable by running
REPORT_DATE='YYYY-MM-DD' make config
OR
REPORT_DATE='YYYY-MM-DD' make rerun-config
OR
REPORT_DATE='YYYY-MM-DD' make rerun-prod

* Fix comment format

Co-authored-by: Katie Gostic <uep6@cdc.gov>

---------

Co-authored-by: Katie Gostic <uep6@cdc.gov>