Poll for config generation completion
Adds a simple shell script that hits the GH API every 15 seconds and
checks the status of the latest workflow run. It assumes that the latest
workflow run in the config generator repo is the run of interest.

I don't touch the time-delay in the `job.py` script but can if we think
it's not needed anymore.