Fiddle with lints on object names
We're getting the dreaded lint on ObjectNameFormats. I think it's fine
so nulling the linter. Not sure why it's only flagging now.

I'm doing it by replacing `object_usage_linter`'s NULL because I think
that one actually should be running. I was very careful in package dev
to handle fixes for that manually and avoiding NSE except where required.