Switch between random walk and GP
This change allows us to switch between the GP and the random walk
options based on the priors specified in the config. It errors if priors
for both the GP and the RW are supplied.

It still needs test coverage and a change in the config generator to
actually supply the needed values.