Move production run earlier (#371)
* Move production run earlier

This change moves the time of the production run up to 11 AM UTC.
This adjustment in timing should reduce the number of hours needed
to wait for runs to complete, requiring less time to maintain
production activities.

This timing is the same as for the HGAM runs, which have been
successful for multiple weeks now.

* Bump NEWS